Make event module a file
